fix(memory): fix import paths and types after provider-runtime refactor
- extensions/memory-core/src/memory/episodic shims: 4→5 levels up (../../../../../)
- packages/memory-host-sdk/src/host/read-file.ts: lazy import EpisodicStore from
  correct path (../../../../src/memory/episodic/store.js)
- compaction-hooks.ts: session-files import → packages/memory-host-sdk path
- compaction-hooks.ts: await createEpisodeEncoder (it is async)
- manager.ts: add EpisodeSearchResult type import, annotate filter/map callbacks
