+++ extensions/memory-core/src/memory/episodic/store.ts
@@ -1,4 +1,4 @@
 // Re-export EpisodicStore from the canonical src/memory/episodic/store.ts.
 // extensions/memory-core/src/memory/manager.ts imports from ./episodic/store.js;
 // this shim makes that import resolvable without duplicating the implementation.
-export { EpisodicStore } from "../../../../src/memory/episodic/store.js";
+export { EpisodicStore } from "../../../../../src/memory/episodic/store.js";
